move HttpStrictTransportSecurityModel namespace one level up to Vanilla\Web
