+++ bootstrap.php
@@ -173,7 +173,7 @@ $dic->setInstance('Garden\Container\Container', $dic)
     ->addCall('setLocaleKey', [ContainerUtils::currentLocale()])
     ->addCall('setCacheBusterKey', [ContainerUtils::cacheBuster()])
 
-    ->rule(\Vanilla\Web\HttpStrictTransportSecurity\HttpStrictTransportSecurityModel::class)
+    ->rule(\Vanilla\Web\HttpStrictTransportSecurityModel::class)
     ->addAlias('HstsModel')
 
     ->rule(\Vanilla\Web\ContentSecurityPolicy\ContentSecurityPolicyModel::class)
