+++ tests/Bootstrap.php
@@ -219,7 +219,7 @@ class Bootstrap {
             ->setShared(true)
             ->addCall('addRoute', ['route' => new \Garden\Container\Reference('@api-v2-route'), 'api-v2'])
 
-            ->rule(\Vanilla\Web\HttpStrictTransportSecurity\HttpStrictTransportSecurityModel::class)
+            ->rule(\Vanilla\Web\HttpStrictTransportSecurityModel::class)
             ->addAlias('HstsModel')
 
             ->rule('@api-v2-route')
